+++ examples/viewer/src/App.jsx
@@ -34,6 +34,7 @@ import {
 } from 'wix-rich-content-plugin-mentions/dist/module.viewer';
 import { fileUploadTypeMapper } from 'wix-rich-content-plugin-file-upload/dist/module.viewer';
 import { giphyTypeMapper } from 'wix-rich-content-plugin-giphy/dist/module.viewer';
+import { buttonTypeMapper } from 'wix-rich-content-plugin-button/dist/module.viewer';
 
 import 'wix-rich-content-common/dist/styles.min.css';
 import 'wix-rich-content-viewer/dist/styles.min.css';
@@ -50,6 +51,7 @@ import 'wix-rich-content-plugin-sound-cloud/dist/styles.min.css';
 import 'wix-rich-content-plugin-map/dist/styles.min.css';
 import 'wix-rich-content-plugin-file-upload/dist/styles.min.css';
 import 'wix-rich-content-plugin-giphy/dist/styles.min.css';
+import 'wix-rich-content-plugin-button/dist/styles.min.css';
 
 import TestData from './TestData/initial-state';
 import styles from './App.scss';
@@ -98,6 +100,7 @@ class App extends Component {
       mapTypeMapper,
       fileUploadTypeMapper,
       giphyTypeMapper,
+      buttonTypeMapper,
     ];
 
     this.config = {